# backend/src/routes/locations.ts
import type { Router } from 'express';
import { Router as createRouter } from 'express';
import {
  createLocation,
  deleteLocation,
  getLocation,
  getLocationHistory,
  listLocations,
  reorderLocation,
  setPrimaryLocation,
  updateWeather,
} from '../db.js';
import {
  SingaporeWeatherClient,
  WeatherProviderError,
  type ForecastArea,
  type WeatherSnapshot,
} from '../weather.js';
import { logger } from '../logger.js';

export interface WeatherClient {
  getCurrentWeather(latitude: number, longitude: number): Promise<WeatherSnapshot>;
  getForecastAreas(): Promise<ForecastArea[]>;
}

interface LocationsRouterOptions {
  weatherClient?: WeatherClient;
}

export function createLocationsRouter(options: LocationsRouterOptions = {}): Router {
  const router: Router = createRouter();
  const weatherClient =
    options.weatherClient ?? new SingaporeWeatherClient({ apiKey: process.env.WEATHER_API_KEY });

  router.get('/locations', async (_request, response, next) => {
    try {
      response.json({ locations: await listLocations() });
    } catch (error) {
      next(error);
    }
  });

  router.get('/forecast-areas', async (_request, response, next) => {
    try {
      response.json({ areas: await weatherClient.getForecastAreas() });
    } catch (error) {
      if (error instanceof WeatherProviderError) {
        response.status(502).json({ detail: 'Forecast areas are unavailable. Please try again.' });
        return;
      }
      next(error);
    }
  });

  router.post('/locations', async (request, response, next) => {
    try {
      const latitude = Number(request.body?.latitude);
      const longitude = Number(request.body?.longitude);

      if (Number.isNaN(latitude) || Number.isNaN(longitude)) {
        response.status(422).json({ detail: 'latitude and longitude are required' });
        return;
      }
      if (!(1.1 <= latitude && latitude <= 1.5 && 103.6 <= longitude && longitude <= 104.1)) {
        response.status(422).json({
          detail: 'Coordinates must be within Singapore (lat 1.1-1.5, lon 103.6-104.1)',
        });
        return;
      }

      const location = await createLocation(latitude, longitude);

      try {
        const snapshot = await weatherClient.getCurrentWeather(
          location.latitude,
          location.longitude,
        );
        const updated = await updateWeather(location.id, snapshot);
        response.status(201).json(updated ?? location);
      } catch (error) {
        if (!(error instanceof WeatherProviderError)) throw error;
        logger.warn(
          { err: error, locationId: location.id },
          'weather refresh failed after location create',
        );
        response.status(201).json(location);
      }
    } catch (error) {
      if (error instanceof Error && error.name === 'DuplicateLocationError') {
        logger.warn({ err: error }, 'duplicate location rejected');
        response.status(409).json({ detail: error.message });
        return;
      }
      next(error);
    }
  });

  router.get('/locations/:locationId', async (request, response, next) => {
    try {
      const location = await getLocation(Number(request.params.locationId));
      if (!location) {
        response.status(404).json({ detail: 'Location not found' });
        return;
      }
      response.json(location);
    } catch (error) {
      next(error);
    }
  });

  router.get('/locations/:locationId/history', async (request, response, next) => {
    try {
      const locationId = Number(request.params.locationId);
      if (!Number.isInteger(locationId) || locationId < 1) {
        response.status(400).json({ detail: 'locationId must be a positive integer' });
        return;
      }

      const requestedLimit = request.query.limit;
      const parsedLimit = requestedLimit === undefined ? 240 : Number(requestedLimit);
      if (!Number.isInteger(parsedLimit) || parsedLimit < 1) {
        response.status(400).json({ detail: 'limit must be a positive integer' });
        return;
      }

      const readings = await getLocationHistory(locationId, Math.min(parsedLimit, 1000));
      if (!readings) {
        response.status(404).json({ detail: 'Location not found' });
        return;
      }

      response.json({ location_id: locationId, readings });
    } catch (error) {
      next(error);
    }
  });

  router.patch('/locations/:locationId/order', async (request, response, next) => {
    try {
      const locationId = Number(request.params.locationId);
      if (!Number.isInteger(locationId) || locationId < 1) {
        response.status(400).json({ detail: 'locationId must be a positive integer' });
        return;
      }

      const direction = request.body?.direction;
      if (direction !== 'up' && direction !== 'down') {
        response.status(400).json({ detail: 'direction must be "up" or "down"' });
        return;
      }

      const result = await reorderLocation(locationId, direction);
      if (result.outcome === 'not_found') {
        response.status(404).json({ detail: 'Location not found' });
        return;
      }
      if (result.outcome === 'primary') {
        response.status(409).json({ detail: 'Primary location cannot be reordered' });
        return;
      }
      if (result.outcome === 'boundary') {
        response.status(409).json({ detail: 'Location is already at the list boundary' });
        return;
      }

      response.json({ locations: result.locations });
    } catch (error) {
      next(error);
    }
  });

  router.post('/locations/:locationId/primary', async (request, response, next) => {
    try {
      const locationId = Number(request.params.locationId);
      if (!Number.isInteger(locationId) || locationId < 1) {
        response.status(400).json({ detail: 'locationId must be a positive integer' });
        return;
      }

      const result = await setPrimaryLocation(locationId);
      if (result.outcome === 'not_found') {
        response.status(404).json({ detail: 'Location not found' });
        return;
      }

      response.json({ locations: result.locations });
    } catch (error) {
      next(error);
    }
  });

  router.delete('/locations/:locationId', async (request, response, next) => {
    try {
      const locationId = Number(request.params.locationId);
      const deleted = Number.isInteger(locationId) && (await deleteLocation(locationId));
      if (!deleted) {
        response.status(404).json({ detail: 'Location not found' });
        return;
      }
      response.status(204).end();
    } catch (error) {
      next(error);
    }
  });

  router.post('/locations/:locationId/refresh', async (request, response, next) => {
    try {
      const locationId = Number(request.params.locationId);
      const location = await getLocation(locationId);
      if (!location) {
        response.status(404).json({ detail: 'Location not found' });
        return;
      }

      const snapshot = await weatherClient.getCurrentWeather(location.latitude, location.longitude);
      const updated = await updateWeather(locationId, snapshot);
      response.json(updated);
    } catch (error) {
      if (error instanceof WeatherProviderError) {
        response.status(502).json({ detail: error.message });
        return;
      }
      next(error);
    }
  });

  return router;
}